refactor(kubernautagent): split server/handler.go into focused files
Extracts the InvestigationResult -> IncidentResponse mapping functions
(mapInvestigationResultToResponse, buildRootCauseAnalysis,
buildSelectedWorkflow, buildAlignmentVerdictResponse,
mapHumanReviewReason, etc.) into handler_response_mapping.go. Reduces
handler.go from 939 to 643 lines (HTTP endpoint handlers only), no
behavioral changes.
